requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.4.0
aiosqlite==0.20.0
annotated-types==0.7.0
APScheduler==3.10.4
attrs==25.3.0
black==24.8.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.141.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
httplib2==0.22.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.7.4
pydantic-settings==2.4.0
pydantic_core==2.18.4
pyparsing==3.2.3
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
ruff==0.5.7
six==1.17.0
SQLAlchemy==2.0.32
typing_extensions==4.14.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1